% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{times2}
\alias{times2}
\title{Times number by 2}
\usage{
times2(x)
}
\arguments{
\item{x}{A numeric or character value}
}
\value{
Output the input number multiplied by 2
}
\description{
Times number by 2
}
\examples{
times2(2)
}
